refactor: moved tracks creation stuff to tracks
